DatalakeBuilder/requirements.txt
boto3
nltk
hazelcast-python-client
pymongo